Changed hostname and fixed bug with bootstrap.sh

+++ bootstrap.sh
@@ -1,5 +1,11 @@
 #!/bin/sh
 
-mkdir -p /etc/puppet/modules
-cd /etc/puppet 
-ln -fs /vagrant/puppet/modules modules
+
+if [ ! -L '/etc/puppet/modules' ]; then
+	echo "=> Creating a symlink to /vagrant/puppet/modules"
+	mkdir -p /etc/puppet
+	cd /etc/puppet 
+	ln -fs /vagrant/puppet/modules modules
+fi 
+
+
